requirements.txt
flask
tensorflow-cpu
matplotlib
numpy
gunicorn